+++ packages/block-editor/src/components/block-list/block.js
@@ -478,16 +478,6 @@ function BlockListBlock( {
 		blockEdit = <div style={ { display: 'none' } }>{ blockEdit }</div>;
 	}
 
-	// Disable reasons:
-	//
-	//  jsx-a11y/mouse-events-have-key-events:
-	//   - onMouseOver is explicitly handling hover effects
-	//
-	//  jsx-a11y/no-static-element-interactions:
-	//   - Each block can be selected by clicking on it
-
-	/* eslint-disable jsx-a11y/mouse-events-have-key-events, jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
-
 	return (
 		<IgnoreNestedEvents
 			id={ blockElementId }
@@ -618,7 +608,6 @@ function BlockListBlock( {
 			) }
 		</IgnoreNestedEvents>
 	);
-	/* eslint-enable jsx-a11y/mouse-events-have-key-events, jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */
 }
 
 const applyWithSelect = withSelect(