+++ packages/eslint-plugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 ### New Features
 
 - New Rule: [`@wordpress/no-unguarded-get-range-at`](https://github.com/WordPress/gutenberg/blob/master/packages/eslint-plugin/docs/rules/no-unguarded-get-range-at.md)
+- Enable `wp` global by default in the `recommended` config.
 
 ## 2.4.0 (2019-08-05)